added a milion changes.. but it's working and stable now, more than ever..
- so.. yeah! :)
