+++ lib/state.py
@@ -1,69 +1,168 @@
-#!/usr/bin/env python3
 
-# Loads and saves state from a JSON file (e.g., state.json).
-
-import json
 import datetime
-import subprocess
+import lib.config as config
+import lib.wifi as wifi
 
 from lib.logger import logger
-from lib.util import from_iso_format_fallback
-from lib.config import Config
-
+from lib.notification import Notification
+from lib.utilities import sync_time
+from lib.gopro import GoPro
 
 class State:
 
     def __init__(self):
-        config = Config()
-        self.config = config
+        self.config = config.global_config
+        self.current_state = config.global_config.state
+        self.wifi = wifi.wifi
+        self.gopro = GoPro()
+        self.notification = Notification()
 
-    def load_state(self):
-        logger.info(f"Loading state. This only happens after a REBOOT. Current state is [{self.config.state}]")
-        result = subprocess.run(["sudo", "uptime"], capture_output=True, text=True)
-        logger.info(f"Uptime: {result.stdout.strip()}")
+    def handle_waiting_state(self):
+        now = datetime.datetime.now()
+        minute = now.minute
+        second = now.second
 
-        try:
-            with open(self.config.state_file_path, "r") as f:
-                loaded_state = json.load(f)
+        if minute in self.config.photo_timer and second < 30:
+            if self.config.last_photo_minute != minute:
+                logger.info("It's time to take a photo. Transitioning to TAKE_PHOTO.")
+                self.current_state = "TAKE_PHOTO"
+            else:
+                logger.debug("Already took a photo this minute. Doing nothing.")
+        # Only 20 sec. here, because it usually takes 25 seconds to complete, and there's no need to do it twice
+        elif minute in self.config.keep_alive_timer and second < 20:
+            self.wifi.keep_alive(True)
+        else:
+            self.wifi.keep_alive(False)
 
-            # Convert string timestamps back to datetime objects
-            for key in ["last_photo_minute", "last_offline_alert_time"]:
-                if key in loaded_state:
-                    loaded_state[key] = from_iso_format_fallback(loaded_state[key])
+    def handle_take_photo_state(self):
+        now = datetime.datetime.now()
+        minute = now.minute
+        second = now.second
 
-            self.config.last_photo_minute = loaded_state.get("last_photo_minute")
-            self.config.last_offline_alert_time = loaded_state.get("last_offline_alert_time")
-            self.config.photo_capture_error_counter = loaded_state.get("photo_capture_error_counter", 0)
-            self.config.error_retries_counter = loaded_state.get("error_retries", 0)
-            self.config.max_error_retries = loaded_state.get("max_error_retries", 5)
-            self.config.restart_counter = loaded_state.get("restart_counter", -1)
-            self.config.sending_alert_every_20_min = loaded_state.get("sending_alert_every_20_min", False)
-
-            logger.info("New stat is loaded and all saved variables are rehydrated.")
-            return loaded_state
-        except FileNotFoundError as e:
-            logger.error(f"Error loading state file: {e}")
-            return {}
-        except Exception as e:
-            logger.error(f"Error loading state: {e}")
-            return {}
-
-    def save_state(self):
-        data = {
-            "last_photo_minute": self.config.last_photo_minute if self.config.last_photo_minute else None,
-            "last_offline_alert_time": from_iso_format_fallback(
-                self.config.last_offline_alert_time.isoformat()) if self.config.last_offline_alert_time else None,
-            "photo_capture_error_counter": self.config.photo_capture_error_counter if self.config.photo_capture_error_counter else 0,
-            "error_retries": self.config.error_retries_counter if self.config.error_retries_counter else 0,
-            "max_error_retries": self.config.max_error_retries if self.config.max_error_retries else 5,
-            "execution_time_seconds": (datetime.datetime.now() - self.config.execution_start_time).total_seconds(),
-            "restart_counter": self.config.restart_counter,
-            "sending_alert_every_20_min": self.config.sending_alert_every_20_min
-        }
+        gopro_ssid = self.config.gopro_config["ssid"]
+        if not self.wifi.ensure_wifi_connected(gopro_ssid):
+            # If we can’t connect to GoPro Wi-Fi, error out
+            logger.error(f"Failed to connect to GoPro Wi-Fi. We are in [{self.current_state}].")
+            return
 
         try:
-            with open(self.config.state_file_path, "w") as f:
-                json.dump(data, f, indent=4)
-            logger.info(f"Saved state to {self.config.state_file_path}.")
+            self.gopro.take_photo()
+            self.config.last_photo_minute = datetime.datetime.now().minute
         except Exception as e:
-            logger.error(f"Failed to save state: {e}")
+            logger.error(f"Error taking photo: {e}. Check logs for more info.")
+            self.config.photo_capture_error_counter += 1
+            self.current_state = "ERROR"
+
+        if (minute == 51 and second > 20) or (minute == 52 and second < 30):
+            self.current_state = "SEND_UPDATE"
+        else:
+            self.current_state = "WAITING"
+
+    def handle_send_update_state(self):
+        """
+        Switch to router Wi-Fi, time sync, send status, save state,
+        then switch back to GoPro Wi-Fi to keep it alive. Return to WAITING when done.
+        """
+        router_ssid = self.config.router_config["ssid"]
+
+        if not self.wifi.ensure_wifi_connected(router_ssid):
+            logger.error(f"Failed to connect to {router_ssid}. ERROR.")
+            self.current_state = "ERROR"
+            return
+
+        sync_time()
+        self.notification.send_status()
+        self.current_state.save_current_configs()
+
+        gopro_ssid = self.config.gopro_config["ssid"]
+        if not self.wifi.ensure_wifi_connected(gopro_ssid):
+            logger.error(f"Failed to reconnect to {gopro_ssid} after sending status. ERROR.")
+            self.current_state = "ERROR"
+            return
+
+        logger.info("Update complete. Returning to WAITING.")
+        self.current_state = "WAITING"
+
+    def handle_error_state(self):
+        """
+          If that fails too many times, we reboot the r-pi.
+        """
+        self.config.error_retries_counter += 1
+        logger.error(f"Error State reached. Retry attempt {self.config.error_retries_counter} / {self.config.max_error_retries}.")
+
+        if self.config.error_retries_counter < self.config.max_error_retries:
+            logger.info("Will attempt to recover by returning to WAITING.")
+            self.current_state = "WAITING"
+        elif self.config.photo_capture_error_counter > 3:
+            logger.error \
+                (f"Too many photo capture errors {self.config.photo_capture_error_counter}. Will attempt to recover by returning to WAITING.")
+            self.current_state = "OFFLINE_ALERT"
+        else:
+            logger.error("Exceeded max error retries. Considering a forced reboot or state reset.")
+            # if more than 5 minutes passed.. there is a good chance GoPro is off, so we can't recover.
+            # todo: add a check to see if GoPro is reachable before rebooting
+            self.current_state = "WAITING"
+            self.config.error_retries_counter = 0
+
+    def handle_offline_alert_state(self):
+        """
+        Repeatedly tries to connect to router Wi-Fi,
+        then attempts GoPro Wi-Fi. If both fail, remains offline.
+
+        - Send one immediate 'GoPro offline' notification if this is the first time.
+        - Then send repeated notifications every 20 minutes if it remains offline (and there's internet).
+        """
+        router_ssid = self.config.router_config["ssid"]
+        gopro_ssid = self.config.gopro_config["ssid"]
+        now = datetime.datetime.now()
+
+        # Make sure the attribute exists
+        if not hasattr(self, "last_offline_alert_time"):
+            # For a new session, pretend we started “offline” a while back
+            # so we can send an alert quickly if needed
+            self.config.last_offline_alert_time = now - datetime.timedelta(minutes=30)
+
+        router_is_up = self.wifi.ensure_wifi_connected(router_ssid)
+        gopro_is_up = False
+
+        if router_is_up:
+            gopro_is_up = self.wifi.ensure_wifi_connected(gopro_ssid)
+
+        if router_is_up and gopro_is_up:
+            # Fully online: reset everything
+            logger.info("Router AND GoPro reconnected. Returning to WAITING.")
+            self.current_state = "WAITING"
+            self.config.last_offline_alert_time = None
+            self.config.sending_alert_every_20_min = False
+        else:
+            # Either router is down, or GoPro is down (or both)
+            if not self.config.last_offline_alert_time:
+                self.config.last_offline_alert_time = now
+
+            time_since_last_offline = now - self.config.last_offline_alert_time
+
+            # If the router is offline, you might restart or do something else here...
+            if not router_is_up:
+                logger.warning("Router is offline, remain in OFFLINE_ALERT.")
+                # self.restart_wifi()
+                # Possibly bail out early, since we can’t reach the GoPro if router is off
+                return
+
+            # Otherwise, router is up but GoPro is offline
+            if time_since_last_offline.total_seconds() >= 1200:  # 20-minute repeat
+                self.notification.send_notification(
+                    title="GoPro OFFLINE!",
+                    message="STILL cannot connect to GoPro. [20 min repeat]. Sent at ..."
+                )
+                self.config.last_offline_alert_time = now
+            else:
+                # If we haven't started the repeating cycle, send that immediate “first offline” alert
+                if not self.config.sending_alert_every_20_min:
+                    self.notification.send_notification(
+                        title="GoPro OFFLINE!",
+                        message="I cannot connect to GoPro. Sending first-time alert at ..."
+                    )
+                    self.config.sending_alert_every_20_min = True
+                    self.config.last_offline_alert_time = now
+
+state_instance = State()
