+++ util/timelapse.service
@@ -20,7 +20,7 @@ StartLimitBurst=3
 Type=simple
 
 # Load configuration and execute script
-ExecStart=/bin/bash -c "config_file='/home/timelapse/config/config.json'; username=$(jq -r .rpi.username $config_file); script_path=$(jq -r .rpi.path $config_file); work_dir=$(jq -r .rpi.work_dir $config_file); . /home/gopro_env/bin/activate && python $script_path"
+ExecStart=/bin/bash -c "config_file='/home/timelapse/config.json'; username=$(jq -r .rpi.username $config_file); script_path=$(jq -r .rpi.path $config_file); work_dir=$(jq -r .rpi.work_dir $config_file); . /home/gopro_env/bin/activate && python $script_path"
 
 Restart=on-failure
 RestartSec=60
